==> R/00_all.qmd <==
---
title: "00_all.qmd"
format: html
editor: visual
author: "Anna Boateng: s175562, Marlies Goedknegt: s220675, Anne Gjelstrup: s194530, Katja Jagd: s185395, Olivia Ronnenberg: s183359"
---

## Part 00: All parts

```{r}
#| label: load packages
#| include: false

library(quarto)
```

```{r}
#| label: rendering
#| include: false

quarto_render(input = "01_load.qmd")
quarto_render(input = "02_clean.qmd")
quarto_render(input = "03_augment.qmd")
quarto_render(input = "04_describe.qmd")
quarto_render(input = "05_analysis_01.qmd")
quarto_render(input = "06_analysis_02.qmd")
```

```{r}
#| label: setting correct path for output 
#| include: false

file.rename(from = "01_load.html",
            to = "../results/01_load.html")
file.rename(from = "01_load_files/",
            to = "../results/01_load_files/")

file.rename(from = "02_clean.html",
            to = "../results/02_clean.html")
file.rename(from = "02_clean_files/",
            to = "../results/02_clean_files/")

file.rename(from = "03_augment.html",
            to = "../results/03_augment.html")
file.rename(from = "03_augment_files/",
            to = "../results/03_augment_files/")

file.rename(from = "04_describe.html",
            to = "../results/04_describe.html")
file.rename(from = "04_describe_files/",
            to = "../results/04_describe_files/")

file.rename(from = "05_analysis_01.html",
            to = "../results/05_analysis_01.html")
file.rename(from = "05_analysis_01_files/",
            to = "../results/05_analysis_01_files/")

file.rename(from = "06_analysis_02.html",
            to = "../results/06_analysis_02.html")
file.rename(from = "06_analysis_02_files/",
            to = "../results/06_analysis_02_files/")
```

{{< include 01_load.qmd >}}

{{< include 02_clean.qmd >}}

{{< include 03_augment.qmd >}}

{{< include 04_describe.qmd >}}

{{< include 05_analysis_01.qmd >}}

{{< include 06_analysis_02.qmd >}}
