==> R/01_load.qmd <==
---
title: "Part 01: Load"
format: html
editor: visual
author: "Anna Boateng: s175562, Marlies Goedknegt: s220675, Anne Gjelstrup: s194530, Katja Jagd: s185395, Olivia Ronnenberg: s183359"
---

The first part of the project workflow involves loading in the data from the original source.

### Import the required packages

```{r}
#| label: load packages
#| message: false

library("tidyverse")
library("readxl")

```

### Load the data

The data, used for this project, is downloaded from https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5580991/. More specifically; supplementary table 1 and 3. These tables contains information on genetic expression data from a study on the influence of disease modifying anti-rheumatic drugs (DMARD) treatment in rheumatoid arthritis (RA) patients.

Data table 1 contain information on co-expression modules. Each column is a group (called module) of co-expressed genes. Each cell belonging to a column contains a gene that is part of this respective module. A total of 20 modules of co-expressed genes is defined. This data frame is therefore referred to as the "coexpression" data frame. The modules are hypothesized to represent biological pathways, but are not assigned to specific ones in the original work. Therefore, the names module 1-20 is retained in this project. The genes in this table relates to data table 3 as they are target molecules of upstream regulators described in the latter mentioned table.

Data table 3 consists of data on several upstream regulators. Each row describes one upstream regulator molecule by characteristics such as the expression level, molecule type and target molecules as mentioned before. Data table 3 exists in two versions. One with data reflecting the RA patients before treatment and one after. The two data frames are identical with respect to variables but not observations, that is the upstream regulators. There is some overlap in the observations but they also differ in which upstream molecules they describe. The data frame before treatment has a total of 849 upstream regulator whereas the data frame after has 595 upstream regulators.

When loading the data, blank fields are replaced with "NA".

```{r}
#| label: load data
#| message: false

coexpression_raw <- read_excel("../data/_raw/pone.0183928.s001.xlsx",
                               na = c(" ",
                                      "NA"))

regulators_before_raw <- read_excel("../data/_raw/pone.0183928.s003.xlsx",
                                    na = c(" ",
                                           "NA"))

regulators_after_raw <- read_excel("../data/_raw/pone.0183928.s003.xlsx",
                                   2,
                                   na = c(" ",
                                          "NA"))

```

### Convert the data files to the correct format

In order to enable data analysis, the original .xlsx files are converted to zipped .tsv and saved in the data folder of the project.

```{r}
#| label: save data

write_tsv(coexpression_raw,
           "../data/01_coexpression_load.tsv.gz",
           col_names = TRUE)

write_tsv(regulators_before_raw,
          "../data/01_regulators_before_load.tsv.gz",
          col_names = TRUE)

write_tsv(regulators_after_raw,
          "../data/01_regulators_after_load.tsv.gz",
          col_names = TRUE)

```
