==> R/03_augment.qmd <==
---
title: "Part 03: Augment"
format: html
editor: visual
author: "Anna Boateng: s175562, Marlies Goedknegt: s220675, Anne Gjelstrup: s194530, Katja Jagd: s185395, Olivia Ronnenberg: s183359"
---

The third part of the project is to augment the cleaned data.

### Import required packages

```{r}
#| label: load packages
#| message: false

library("tidyverse")
```

### Accessing relevant functions

```{r}
#| label: use functions from file

source("99_proj_func.R")
```

### Load in cleaned data files

```{r}
#| label: load data

coexpression_aug <- read_tsv("../data/02_coexpression_clean.tsv.gz",
                             show_col_types = FALSE)

expr_intersect_aug <- read_tsv("../data/02_expr_intersect_clean.tsv.gz",
                               show_col_types = FALSE)

target_intersect_aug <- read_tsv("../data/02_target_intersect_clean.tsv.gz",
                                 show_col_types = FALSE)

expr_before_aug <- read_tsv("../data/02_expr_before_clean.tsv.gz",
                            show_col_types = FALSE)

expr_after_aug <- read_tsv("../data/02_expr_after_clean.tsv.gz",
                           show_col_types = FALSE)
```

### Augment the data frames

#### Augment data frames "expr_before" and "expr_after"

The function used below creates a column that indicates whether a data point is significant or not. Furthermore, a -log10 transformation of the p-values is added.

```{r}
#| label: classifying significance

expr_before_aug <- significant_label(expr_before_aug,
                                     p_value_before,
                                     expr_log_ratio_before)

expr_after_aug <- significant_label(expr_after_aug,
                                    p_value_after,
                                    expr_log_ratio_after) 
```

#### Augment data frame "expr_intersect"

A variable that shows the level of change in gene expression for the regulators is added.

```{r}
#| label: add change in expr column 

expr_intersect_aug <- expr_intersect_aug |>
  mutate(change_in_expr = expr_log_ratio_after - expr_log_ratio_before) |> 
  relocate("change_in_expr",
           .after = "expr_log_ratio_after")
```

A variable indicating whether a regulator was up- or down-regulated both before and after treatment is added.

```{r}
#| label: add column showing up- or downregulation 

expr_intersect_aug <- expr_intersect_aug |>
  mutate(regulation = case_when(change_in_expr > 0 ~ "Upregulated",
                                change_in_expr < 0 ~ "Downregulated")) |> 
  relocate("regulation",
           .after = "change_in_expr")
```

#### Augment data frame "target_intersect"

A variable called "change_in_n_molecules" showing the difference in target molecules after treatment compared to before treatment was added for each regulator. Along with a classification variable called "change_class", which determines if the change in molecules leads to an increase, decrease, or no change of target molecules.

```{r}
#| label: change classification

target_intersect_aug <- target_intersect_aug |> 
  mutate(mol_change = n_target_after - n_target_before,
         change_class = case_when(mol_change > 0 ~ "Increase",
                                  mol_change < 0 ~  "Decrease",
                                  mol_change == 0 ~ "Unchaged")) |> 
  relocate(mol_change, 
           change_class,
           .after = n_target_after) |> 
  arrange(desc(upstream_regulator)) 
```

### Saving the augmented data frames in new files

Lastly, the augmented data frames were saved as zipped .tsv files.

```{r}
#| label: save data

write_tsv(coexpression_aug,
          "../data/03_coexpression_aug.tsv.gz")

write_tsv(expr_intersect_aug,
          "../data/03_expr_intersect_aug.tsv.gz")

write_tsv(target_intersect_aug,
          "../data/03_target_intersect_aug.tsv.gz")

write_tsv(expr_before_aug,
          "../data/03_expr_before_aug.tsv.gz")

write_tsv(expr_after_aug,
          "../data/03_expr_after_aug.tsv.gz")
```
