==> R/05_analysis_01.qmd <==
---
title: "Part 05: analysis 1"
format: html
editor: visual
author: "Anna Boateng: s175562, Marlies Goedknegt: s220675, Anne Gjelstrup: s194530, Katja Jagd: s185395, Olivia Ronnenberg: s183359"
---

The fifth part of the project is the first analysis, where genetic expression levels of the upstream regulators are analyzed. Here it is also investigated how these are influenced by DMARD treatment.

### Load the required packages

```{r}
#| label: load packages
#| message: false

library("tidyverse")
library("ggrepel")

```

### Load the needed .tsv.gz files

```{r}
#| label: load files

expr_before_analysis <- read_tsv("../data/03_expr_before_aug.tsv.gz",
                                 show_col_types = FALSE)

expr_after_analysis <- read_tsv("../data/03_expr_after_aug.tsv.gz",
                                show_col_types = FALSE)

expr_intersect_analysis <- read_tsv("../data/03_expr_intersect_aug.tsv.gz",
                                    show_col_types = FALSE)

```

### Generate scatter plots of the gene expression

To depict which upstream regulators are expressed significantly different from the healthy control group, two scatter plots are made. One plot comparing patients before treatment, and the other comparing patients after treatment to the control group. The significant upstream regulators are depicted in turquoise and insignificant in red.

```{r}
#| label: volcano plot before treatment 
#| warning: false 

expr_before_analysis <- expr_before_analysis |>
  mutate(lbl = case_when(is_significant == "yes" ~ upstream_regulator,
                         is_significant == "no" ~ ""))

plot_before <- expr_before_analysis |>
  ggplot(aes(x = expr_log_ratio_before,
             y = `-log10_p_value`,
             colour = is_significant,
             label = lbl)) + 
  geom_text_repel(size = 3,
                  max.overlaps = 25) + 
  geom_point() +
  xlim(-5, 5) + 
  geom_hline(yintercept = -log10(0.05), linetype = "dotted", color = "black") + 
  geom_vline(xintercept = c(-1.5, 1.5), linetype = "dotted", color = "black") +
  theme_minimal(base_size = 12) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5),
    legend.position = "bottom") +
  guides(colour = guide_legend(override.aes = list(size = 6))) +
  labs(
    x = "Expression log ratio",
    y = "-log10(p)",
    color = "Significant",
    title = "Genetic expression in RA patients",
    subtitle = "Before DMARD treatment",
    caption = "Data from DOI: 10.1371/journal.pone.0183928")

plot_before

ggsave(filename = "05_expr_plot_before.png",
       plot = last_plot(),
       device = "png",
       path = "../results", 
       scale = 1,
       limitsize = TRUE)

```

```{r}
#| label: volcano plot after treatment 
#| warning: false 

expr_after_analysis <- expr_after_analysis |>
  mutate(lbl = case_when(is_significant == "yes" ~ upstream_regulator,
                         is_significant == "no" ~ ""))

plot_after <- expr_after_analysis |>
  ggplot(aes(x = expr_log_ratio_after,
             y = `-log10_p_value`,
             colour = is_significant,
            label = lbl)) + 
  geom_text_repel(size = 3,
                  max.overlaps = 25) + 
  geom_point() +
  xlim(-5, 5) + 
  geom_hline(yintercept = -log10(0.05), linetype = "dotted", color = "black") + 
  geom_vline(xintercept = c(-1.5, 1.5), linetype = "dotted", color = "black") +
  theme_minimal(base_size = 12) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5),
    legend.position = "bottom") +
  guides(colour = guide_legend(override.aes = list(size = 6))) +
  labs(
    x = "Expression log ratio",
    y = "-log10(p)",
    color = "Significant",
    title = "Genetic expression in RA patients",
    subtitle = "After DMARD treatment",
    caption = "Data from DOI: 10.1371/journal.pone.0183928")

plot_after

ggsave(filename = "05_expr_plot_after.png",
       plot = last_plot(),
       device = "png",
       path = "../results",
       scale = 1,
       limitsize = TRUE)

```

The two scatter plots show that there are significantly differential expressed upstream regulators, both before and after treatment. The criteria for being significant was defined by a p-value \> -log10(0.05) and a log-ratio gene expression between -1.5 and 1.5, indicated by the dotted black lines.

Both unique and common significant upstream regulators are found in both plots. One example of an upstream regulator that both plots contain could be TGFB1, found in the top left corner in both plots. All the common upstream regulators will be further analysed to answer which upstream regulators are influenced by the treatment.

#### Generate a comparison plot

A plot is generated, depicting the influence of DMARD treatment on gene expression. It shows the genetic change in expression before and after treatment along the x-axis. Upstream regulators depicted further from the intercept at x = 0 indicate a larger influence from treatment.

```{r}
#| label: comparison plot
#| warning: false

comparison_plot <- expr_intersect_analysis |> 
  ggplot(aes(x = change_in_expr,
             y = upstream_regulator,
             color = regulation)) +
  geom_point() +
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +
  theme_minimal(base_size = 12) +
  theme(
    axis.text.y=element_text(size=6),
    plot.title = element_text(hjust = 0.5),
    plot.subtitle = element_text(hjust = 0.5),
    legend.position = "bottom") +
  guides(colour = guide_legend(override.aes = list(size = 6))) +
  labs(
    x = "Change in gene expression",
    y = "Upstream Regulator",
    color = "Gene regulation",
    title = "Changes in genetic expression due to treament",
    caption = "Data from DOI: 10.1371/journal.pone.0183928")

comparison_plot

ggsave(filename = "05_comparison_plot.png",
       plot = last_plot(),
       device = "png",
       path = "../results",
       scale = 1,
       limitsize = TRUE)

```

It can be seen from the plot that some upstream regulators are more affected by treatment than others. The following upstream regulators IGF2, LIPE, MLK2, OLR1, PTPRC, RNASE2, S100A9, and TLR2 seems to be most affected by the treatment. They will be further analyzed in "06_analysis_02".
