==> R/06_analysis_02.qmd <==
---
title: "Part 06: analysis 2"
format: html
editor: visual
author: "Anna Boateng: s175562, Marlies Goedknegt: s220675, Anne Gjelstrup: s194530, Katja Jagd: s185395, Olivia Ronnenberg: s183359"
---

The sixth part of the project is the second analysis, where the change in the number of target molecules before and after treatment is analyzed. Furthermore, a deeper analysis of three interesting regulators is performed.

### Import required packages

```{r}
#| label: load packages
#| message: false

library("tidyverse")
```

### Accessing relevant functions

```{r}
#| label: use functions from file

source("99_proj_func.R")
```

### Load the needed .tsv.gz files

```{r}
#| label: load augmented data files

coexpression_analysis <- read_tsv("../data/03_coexpression_aug.tsv.gz",
                                  show_col_types = FALSE)

target_intersect_analysis <- read_tsv("../data/03_target_intersect_aug.tsv.gz",
                                      show_col_types = FALSE)
```

### Generate bar plot of number of target molecules before and after treatment

The upstream regulators each influence several target molecules. To investigate whether this influence was affected by the treatment, a plot was created to indicate the change of the number of target molecules before and after treatment.

```{r}
#| label: generate target molecules plot
#| message: false

targets_plot <- target_intersect_analysis |> 
  ggplot(aes(y = upstream_regulator,
             x = mol_change,
             fill = change_class)) +
  geom_col(color  = "grey", 
           alpha = 0.5) +
  scale_fill_manual(values = c("Decrease" = "#F8766D",
                             "Increase" = "#00BFC4")) +
  theme_minimal(base_size = 12) +
  theme(axis.text.y = element_text(size = 7),
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5)) +
  labs(y = "Upstream regulator",
       x = "Change in number of target molecules",
       title = "Change in the number of target molecules",
       subtitle = "Comparing after treatment to before",
       caption = "Data from DOI: 10.1371/journal.pone.0183928") + 
  scale_x_continuous(breaks = c(-15,-10, -5, 0, 5, 10, 15)) +
  coord_fixed(ratio = 4/3) + 
  geom_vline(xintercept = 0,
             alpha = 0.5,
             color = "grey")

targets_plot

ggsave(filename = "06_targets_plot.png",
       plot = last_plot(),
       device = "png",
       path = "../results",
       scale = 1,
       limitsize = TRUE)


```

The plot indicates that the number of target molecules, influenced by each upstream regulator, changes after treatment. There are cases of both increasing, decreasing and unchanged number of target molecules. This highlights the shift in the behavior of the upstream regulators after treatment, which is relevant information when assessing the effect of the treatment, its possible benefits and side effects.

### Biological context of the target molecules

The analysis of the expression data, together with the target molecule analysis above, revealed the following interesting regulators: LIPE, OLR1, and S100A9. They all have a significant change in expression and in the number of target molecules. These regulators could be potential drug targets.

As mentioned, the modules in the co-expression data set represent possible biological pathways. Therefore a further analysis was performed to look into which modules the target molecules of the candidates regulators belonged to.

By using the function "module_counter()" on each candidate it was possible to connect the information from the co-expression modules with the target molecules of the regulators. The function counts how many target molecules belongs to each module.

```{r}
#| label: define the selected molecule candidates

candidate_1 <- module_counter(target_intersect_analysis, 
                              "LIPE")
candidate_2 <- module_counter(target_intersect_analysis, 
                              "OLR1")
candidate_3 <- module_counter(target_intersect_analysis, 
                              "S100A9")
```

The results from the three candidates was merged into a small data set called "module_analysis".

```{r}
#| label: create data set for the module analysis

module_analysis <- candidate_1 |> 
  full_join(candidate_2,
            by = 'modules') |> 
  full_join(candidate_3,
            by = 'modules') |> 
  mutate_if(is.numeric,
            replace_na,
            0) |> 
  pivot_longer(cols = -modules,
               names_to = "candidate",
               values_to = "counts") |> 
  arrange(as.integer(str_extract(modules, 
                                 "\\d+")))
```

The result of the module analysis is visualized in the bar plot below.

```{r}
#| label: generate module analysis plot
#| message: false

modules_plot <- module_analysis |>
  ggplot(aes(x = fct_inorder(modules), 
             y = counts, 
             fill = candidate)) +
  geom_bar(stat = "identity",
           color = "grey",
           alpha = 0.5,
           position = position_dodge(preserve = "single"),
           ) +
  geom_hline(yintercept = 0) + 
  theme_minimal(base_size = 12) +
  theme(panel.grid.major.x = element_blank(),
        axis.text.x = element_text(angle = 20,
                                   hjust = 1,
                                   vjust = 1),
        legend.title=element_blank(),
        plot.title = element_text(hjust = 0.5)) +
  labs(x = "",
       y = "Number of target molecules",
       title = "Involvement of upstream regulators in biological pathways (modules)",
       subtitle = "Before treatment ",
       caption = "Data from DOI: 10.1371/journal.pone.0183928") +
  scale_fill_manual(values = c( "#F8766D",
                                "#ded031",
                                "#00BFC4")) 

modules_plot

ggsave(filename = "06_modules_plot.png",
       plot = last_plot(),
       device = "png",
       path = "../results",
       scale = 1,
       limitsize = TRUE)
               
                    
```

The plot shows how the three regulators influence specific biological pathways (modules). It is clear from the results that both LIPE, OLR1, and S100A9 have a significant number of target molecules in several modules. Particularly, module 2, 6, and 7 are influenced by these upstream regulators. This could imply that targeting LIPE, OLR3, and S100A9 could have a significant effect in RA treatment.
